pyproject.toml
[project]
name = "var-exp"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "openai>=1.40.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0",
    "scipy>=1.13.0",
    "seaborn>=0.13.0",
    "sentence-transformers>=3.0.0",
    "tqdm>=4.66.0",
]
